+++ test/src/test_type_storage.cpp
@@ -112,3 +112,23 @@ TEST(TestTypeStorage, baget_basetype)
 
     EXPECT_EQ(a.m_c, b.m_c);
 }
+
+TEST(TestTypeStorage, get_baget_const)
+{
+    const std::tuple<std::stringstream,
+               std::string, std::vector<char>, std::mutex, uint64_t,
+               special_a> tup;
+
+    const auto& ba = type_storage::baget<base>(tup);
+    (void) ba;
+
+    const auto& sa = type_storage::get<special_a>(tup);
+    (void) sa;
+
+    // Ignores const statement when searching
+    const auto& bac = type_storage::baget<const base>(tup);
+    (void) bac;
+
+    const auto& sac = type_storage::get<const special_a>(tup);
+    (void) sac;
+}