+++ NEWS.rst
@@ -6,7 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Now ignoring ``const`` and ``volatile`` qualifiers on types.
+* Patch: Fixed bug in const version of ``type_storage::baget()``
 
 2.1.0
 -----